# Optimize keyscanner for 'avr' native keyboards