--- a/src/kaleidoscope/hardware/ATMegaKeyboard.cpp
+++ b/src/kaleidoscope/hardware/ATMegaKeyboard.cpp
@@ -40,11 +40,11 @@ void ATMegaKeyboard::setup(void) {
     OUTPUT_HIGH(KeyboardHardware.matrix_row_pins[i]);
   }
 
-  /* Set up Timer1 for 500usec */
+  /* Set up Timer1 for 1700usec */
   TCCR1B = _BV(WGM13);
   TCCR1A = 0;
 
-  const uint32_t cycles = (F_CPU / 2000000) * 500;
+  const uint32_t cycles = (F_CPU / 2000000) * 1700;
 
   ICR1 = cycles;
   TCCR1B = _BV(WGM13) | _BV(CS10);