--- a/src/kaleidoscope/hardware/ez/ErgoDox.cpp
+++ b/src/kaleidoscope/hardware/ez/ErgoDox.cpp
@@ -67,11 +67,11 @@ void ErgoDox::setup(void) {
   setStatusLEDBrightness(2, 15);
   setStatusLEDBrightness(3, 15);
 
-  /* Set up Timer1 for 500usec */
+  /* Set up Timer1 for 1700usec */
   TCCR1B = _BV(WGM13);
   TCCR1A = 0;
 
-  const uint32_t cycles = (F_CPU / 2000000) * 500;
+  const uint32_t cycles = (F_CPU / 2000000) * 1700;
 
   ICR1 = cycles;
   TCCR1B = _BV(WGM13) | _BV(CS10);