--- a/src/kaleidoscope/hardware/olkb/Planck.cpp
+++ b/src/kaleidoscope/hardware/olkb/Planck.cpp
@@ -18,6 +18,7 @@
 #ifdef ARDUINO_AVR_PLANCK
 
 #include <Kaleidoscope.h>
+#include <avr/wdt.h>
 
 namespace kaleidoscope {
 namespace hardware {