--- a/src/kaleidoscope/hardware/technomancy/Atreus.cpp
+++ b/src/kaleidoscope/hardware/technomancy/Atreus.cpp
@@ -27,6 +27,7 @@
 #ifdef ARDUINO_AVR_ATREUS
 
 #include <Kaleidoscope.h>
+#include <avr/wdt.h>
 
 namespace kaleidoscope {
 namespace hardware {