@algernon - Please have a look at this branch. I believe it's significantly faster than your handrolled atreus keyscanner. (And fixes a serious bug that was masked by having a really slow scanner ;)