@algernon 2018-11-21
Loving it. Code size is bigger (by about 800 bytes), but cycle times went from 0.64ms to 0.16ms. That alone is worth the extra size in my book, especially when paired with a serious bug fixed.

I pushed a few ErgoDox changes that apply similar fixes & optimizations there. The speedup in that case isn't that big (yet): from 2.2ms to 1.6ms using my own sketch, at the cost of ~800 bytes of PROGMEM. There's room to improve in the ErgoDox case, I think, but this is already nice enough.